% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_raster}
\alias{assign_raster}
\title{assign raster}
\usage{
assign_raster(ras, vals)
}
\arguments{
\item{ras}{is a raster}

\item{vals}{are vals for raster}
}
\value{
raster `ras` with values `vals`
}
\description{
wrapper for assigning vals to raster
}
