% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{before_and_after}
\alias{before_and_after}
\title{before and after plots}
\usage{
before_and_after(chng_path, year1_path, year2_path, img_path)
}
\arguments{
\item{chng_path}{is a path for directory of change rasters output by `pred_change()`}

\item{year1_path}{is a character string path for directory of year 1 rasters}

\item{year2_path}{is a character string path for directory of year 2 rasters}

\item{img_path}{is a character string path of directory to output results}
}
\value{
plots to the working directory showing before and after
}
\description{
for a set of rasters, print images for year 1 and year 2
}
