% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bisecter}
\alias{bisecter}
\title{BISECT EXTENT}
\usage{
bisecter(ext, vertical = TRUE)
}
\arguments{
\item{ext}{is an object of class \code{extent}}

\item{logical}{indictor of vertical or horizontal bisection, default TRUE}
}
\value{
a list of the extents of the bisections
}
\description{
Given an extent \code{ext}, returns the extents of the bisections in a list.
}
