% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_change_table}
\alias{build_change_table}
\title{Build Cover Change Table}
\usage{
build_change_table(csv, year1, year2, polys, lots, permits)
}
\arguments{
\item{csv}{is a csv file with cols(sample_id, year, type, obs(1:50))}

\item{year1}{is a number matching a value in csv$year}

\item{year2}{is a number matching a value in csv$year}

\item{polys}{is a spatial polygons object (sampling boxes)}

\item{lots}{is a spatial polygons object (riparian taxlots)}

\item{permits}{is a csv file with cols(maptaxlot, record_no)}
}
\value{
prints a csv file with cols(id, change, maptaxlot, permits) to the working directory
}
\description{
Build Cover Change Table
}
