% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_report}
\alias{change_report}
\title{Cover Change Report}
\usage{
change_report(csv, year1, year2, title = "cover_change.png",
  heading = "\% Cover Change", leg_pos = "topleft",
  y_lab = "No. of Sample Sites with Cover Change of X")
}
\arguments{
\item{year1}{is a matrix with cols (sample_id,year,type,results(1:50))}

\item{year2}{is a matrix same format as year1}
}
\value{
prints a graph of mean cover change into the working directory
}
\description{
Prints a graph of cover change.
}
