% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{color_array}
\alias{color_array}
\title{color_array}
\usage{
color_array(img)
}
\arguments{
\item{img}{is a raster stack of 4 bands rgbn}
}
\value{
an array of summary color statistics
}
\description{
extracts color statistics from raster image
}
