% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{color_array_3band}
\alias{color_array_3band}
\title{color_array_3band}
\usage{
color_array_3band(img)
}
\arguments{
\item{img}{is a raster stack of 3 bands rgb}
}
\value{
an array of summary color statistics
}
\description{
extracts color statistics from raster image
}
