% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_array_set}
\alias{color_array_set}
\title{color_array_set}
\usage{
color_array_set(polys, in_path)
}
\arguments{
\item{polys}{is an sf object}

\item{in_path}{is the directory of raster}
}
\value{
color array data.table
}
\description{
color array set from polys for model calibration.
}
