% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comp_angle}
\alias{comp_angle}
\title{comp_angle}
\usage{
comp_angle(bear1, bear2)
}
\arguments{
\item{bear1}{is a bearing}

\item{bear2}{is a bearing}
}
\value{
difference in bearing
}
\description{
Compare angles of two vectors.
}
\seealso{
orient
}
