% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corners}
\alias{corners}
\title{corners}
\usage{
corners(mat, dist)
}
\arguments{
\item{mat}{is a matrix with cols c(x,y)}

\item{dist}{is the size of riparian corridor}
}
\value{
a list with elements [buffer, corners]
}
\description{
Find the four corners of a sampling box.
}
\seealso{
sample_streams
}
