% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{draw_bank}
\alias{draw_bank}
\title{draw bank}
\usage{
draw_bank(mat, box, bank = "L", crs)
}
\arguments{
\item{mat}{is a buffer matrix output by get_buffer}

\item{box}{is a vector of corner row ids output by get_buffer}

\item{bank}{is 'L' for left bank and 'R' for right bank}
}
\value{
right or left bank segment of sampling box points as matrix
}
\description{
Takes output from get_buffer and draws the left and right bank points for
sampling box.
}
\seealso{
get_buffer
}
