% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ext_box}
\alias{ext_box}
\title{extent box}
\usage{
ext_box(ras)
}
\arguments{
\item{ras}{is a spatial object (raster)}
}
\value{
a Spatial Polygon drawn around the extent of `ras`
}
\description{
produce polygon representation of extent
}
\seealso{
before_and_after
}
