% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_extent}
\alias{fill_extent}
\title{MERGE RASTERS TO FILL EXTENT}
\usage{
fill_extent(poly, dir, band)
}
\arguments{
\item{poly}{is a \code{SpatialPolygons} object}

\item{dir}{is a directory filepath character string}

\item{band}{is the band to extract from multiband imagery}
}
\value{
merged rasters from \code{dir} over extent of \code{poly}
}
\description{
Given a \code{SpatialPolygons} object and a directory filepath character string,
returns merged rasters from \code{dir} over extent of \code{poly}.
}
