% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{get_bear}
\alias{get_bear}
\title{get bearing}
\usage{
get_bear(coords)
}
\arguments{
\item{coords}{is a matrix with cols (x,y) and rows(1:2)}
}
\value{
the angle of the bearing between the points in `coords` in degrees
}
\description{
Given a matrix of two coordinate pairs, returns the bearing of the angle
between the two points in degrees.
}
