% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{get_buffer}
\alias{get_buffer}
\title{get buffer}
\usage{
get_buffer(mat, crs)
}
\arguments{
\item{mat}{is a centerline matrix}
}
\value{
a list containing the buffer matrix and vector of box corner row ids
}
\description{
Given the matrix output from the centerline function, generates the riparian
buffer, adds the four corners of the sampling box explicitly to the buffer matrix,
and returns the buffer and a vector of corner row ids in a list.
}
\seealso{
centerline
}
