% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols_4band}
\alias{get_cols_4band}
\title{Get colors from 4 band raster}
\usage{
get_cols_4band(poly, dir)
}
\arguments{
\item{poly}{is the sampling box}

\item{dir}{is the path to the rasters}
}
\value{
a dt with cols (ndvi, red, grn, blu)
}
\description{
Returns data.table of mean ndvi and rgb values for a sampling box
}
