% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_crs}
\alias{get_crs}
\title{Get CRS}
\usage{
get_crs(dir)
}
\arguments{
\item{dir}{is a path to a directory of rasters}
}
\value{
the CRS character string
}
\description{
Get a reference CRS from rasters in a directory
}
