% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_palette}
\alias{get_palette}
\title{get_palette}
\usage{
get_palette(x, a = 0.33)
}
\arguments{
\item{x}{is a vector of color names or number}

\item{a}{is the alpha transparency (0,1)}
}
\value{
a palette of rgb values
}
\description{
get a palette of colors at chosen transparency
colors are coral, hardwood, gold, forest, sky, ocean
violet, rose, crimson
choose a palette with a vector of names
get a random palette length `x` if `x` is a number
}
