% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ras}
\alias{get_ras}
\title{get_ras}
\usage{
get_ras(path = getwd())
}
\arguments{
\item{path}{is a character vector specifying a directory of files}
}
\value{
a character vector of filenames of rasters in `path`
}
\description{
get filenames of rasters in `path`
}
