% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rasters}
\alias{get_rasters}
\title{get raster names}
\usage{
get_rasters(path = getwd())
}
\arguments{
\item{path}{is a character string specifying a file path to a directory of rasters}
}
\value{
a character vector of raster file names
}
\description{
scan a path directory for raster file names, includes .tif, .hdr
}
