% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_extent}
\alias{in_extent}
\title{IN EXTENT LOGICAL TEST}
\usage{
in_extent(pt, so, is_in = FALSE)
}
\arguments{
\item{pt}{is a coordinate vector or extent class object}

\item{so}{is a spatial object}

\item{is_in}{is a logical binary set to FALSE}
}
\value{
TRUE if pt is within extent of ras, else FALSE
}
\description{
Tests whether the point (or extent) \code{pt} falls entirely within the extent of a raster \code{ras},
returning TRUE or FALSE.
}
