% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_pt}
\alias{insert_pt}
\title{insert_pt}
\usage{
insert_pt(pt, mat)
}
\arguments{
\item{pt}{is the point to add}

\item{mat}{is the line matrix}
}
\value{
list of new matrix and id of pt on mat
}
\description{
Add point to line matrix.
}
