% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inspect_samples}
\alias{inspect_samples}
\title{inspect_samples}
\usage{
inspect_samples(
  samples,
  in_path,
  out_path = getwd(),
  permits = permits_13to18,
  lots = sf_lots
)
}
\arguments{
\item{samples}{is a list of surveying boxes produced by `sample_streams`}

\item{in_path}{is a character vector delineating file path to raster imagery.}

\item{out_path}{is a character vector delineating a file path to print png output.}

\item{permits}{is a permit list with cols c(1, 5) = c(record_no, maptaxlot)}

\item{lots}{is a multipolygon sf object representing maptaxlots in the riparian corridor.}
}
\value{
Prints pngs of maptaxlots covered by sampling boxes for inspection of sample quality.
}
\description{
Prints plots showing the maptaxlots covered by a sampling box.
}
