% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{junction}
\alias{junction}
\title{junction}
\usage{
junction(id, pt, strm_cords, thresh)
}
\arguments{
\item{id}{is the id number of the stream}

\item{pt}{is the current position}

\item{strm_cords}{is a matrix of stream geometry}

\item{thresh}{is the connection threshhold}
}
\value{
matrix of adjacent stream coordinates
}
\description{
Finds adjacent stream lines.
}
