% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{logindex}
\alias{logindex}
\title{logical index}
\usage{
logindex(bool, ids = 0)
}
\arguments{
\item{bool}{is a vector of logical operators}
}
\value{
the indices of true elements along the vector `log`
}
\description{
Given a vector of logical booleans, returns the indices of true elements
}
