% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_lot}
\alias{lookup_lot}
\title{Look up MapTaxLot`}
\usage{
lookup_lot(poly, lots)
}
\arguments{
\item{poly}{is a polygon (sampling box)}

\item{lots}{is an spdf of polygons (bc taxlots)}
}
\value{
the MapTaxlot of the centroid of the input polygon
}
\description{
Given a polygon, returns the MapTaxlot number of the centroid.
}
