% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_permits}
\alias{lookup_permits}
\title{Look up permits}
\usage{
lookup_permits(so, lots, permits)
}
\arguments{
\item{so}{is a spatial polygons object (sampling boxes)}

\item{lots}{is a spatial polygons object (taxlots)}

\item{permits}{is a csv file with cols(maptaxlot, record_no)}
}
\value{
a character vector listing permit record numbers for taxlots in polys
}
\description{
Look up permits
}
