% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_line}
\alias{make_line}
\title{Convert coordinates to SpatialLines}
\usage{
make_line(coords, crs_ref)
}
\arguments{
\item{coords}{is a matrix with cols(x,y)}

\item{crs_ref}{is a crs character string}
}
\value{
a SpatialLines object
}
\description{
Wrapper for sp workflow of nesting spatial objects in lists like russian dolls.
}
