% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_crs}
\alias{match_crs}
\title{MATCH CRS}
\usage{
match_crs(polys, crs_ref)
}
\arguments{
\item{polys}{is a spatial polygons objects}

\item{crs_ref}{is a character string representing the crs}
}
\value{
a list of spatial polygons projected in the common crs
}
\description{
Make sure your extents overlap by converting polygons
into a common reference crs.
}
