% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orient}
\alias{orient}
\title{orient}
\usage{
orient(pt, mat, bear)
}
\arguments{
\item{pt}{is the current position}

\item{mat}{is the line matrix to traverse}

\item{bear}{is the bearing of travel}
}
\value{
matrix leading from pt in direction of travel
}
\description{
Orient a matrix for forward traverse.
}
