% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad}
\alias{pad}
\title{pad}
\usage{
pad(so, pad = 0.1)
}
\arguments{
\item{so}{is an sf object}

\item{pad}{is the factor of padding}
}
\value{
an sf boundary box
}
\description{
pad sf object and return extent
}
\seealso{
fill_extent
}
