% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_change}
\alias{plot_change}
\title{Plot Cover Change}
\usage{
plot_change(csv, year1, year2, title = "cover_change.png",
  heading = " ", leg_pos = "topleft",
  y_lab = "No. of Sample Sites with Cover Change of X")
}
\arguments{
\item{csv}{is a csv with cols (id, year, type, results(1:50))}

\item{year1}{is an integer matching a value in csv$year}

\item{year2}{is an integer matching a value in csv$year}

\item{title}{is the character vector to name the output file (png)}

\item{heading}{is the character vector used as the main title of the plot}

\item{leg_pos}{is the legend position argument (character)}

\item{y_lab}{is the label for the y axis (character)}
}
\value{
prints a graph of mean cover change into the working directory
}
\description{
Prints a graph of cover change.
}
