% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cover}
\alias{plot_cover}
\title{Plot Cover Extent}
\usage{
plot_cover(csv, title = "cover_change.png",
  heading = "\% Cover Change", leg_pos = "topleft",
  y_lab = "Proportion of Samples at or below X\% Cover")
}
\arguments{
\item{csv}{is a csv with cols (id, year, type, results(1:50))}

\item{title}{is the character vector to name the output file (png)}

\item{heading}{is the character vector used as the main title of the plot}

\item{leg_pos}{is the legend position argument (character)}

\item{y_lab}{is the label for the y axis (character)}
}
\value{
prints a graph of mean cover extent into the working directory
}
\description{
Prints a graph of cover extent.
}
