% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pred_cover}
\alias{plot_pred_cover}
\title{plot predicted cover}
\usage{
plot_pred_cover(poly, path, title = "pred_change.png")
}
\arguments{
\item{poly}{is a spatial polygons object (taxlot or sampling box)}

\item{path}{is a character string indicating the path to a directory to rasters}

\item{title}{is the name of the png plot printed to the working directory}
}
\value{
a raster plot of predicted cover masked by `poly`
}
\description{
Predicts cover extent within a polygon and returns the results as a raster plot
}
