% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores}
\alias{plot_scores}
\title{plot_scores}
\usage{
plot_scores(
  in_path,
  out_path,
  polys = samples,
  scores = samples2018,
  title = "samples_"
)
}
\arguments{
\item{in_path}{is a character vector containing the path to the orthoimagery}

\item{out_path}{is a character vector specifying the output directory for the plots}

\item{polys}{is a SpatialPolygonsDataFrame object (the sample boxes shapefile)}

\item{scores}{is a data.table recording cover score observations}
}
\value{
prints rgb plots of `polys` to `out_path` with scores indicated by color
}
\description{
Superimposes sampling boxes over orthographic data.
Prints cover scores over each box.  
`scores` matches output format of plot_samples().  
`scores` must have same number of rows as `polys`, in the same order.
}
