% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{poly_to_poly}
\alias{poly_to_poly}
\title{poly_to_poly}
\usage{
poly_to_poly(poly_list)
}
\arguments{
\item{poly_list}{is a list of SpatialPolygonDataFrame objects}
}
\value{
a SpatialPolygonDataFrame with elements from `poly_list`
}
\description{
extracts list of polygons from each spatial object
add polygons to master list
turn master list into new SpatialPolygonDataFrame object
}
