% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_change}
\alias{pred_change}
\title{predict change}
\usage{
pred_change(year1_path, year2_path, out_path)
}
\arguments{
\item{year1_path}{is a character string indicating path to year 1 rasters}

\item{year2_path}{is a character string indicating path to year 2 rasters}

\item{out_path}{is a character string indicating path to output change rasters}
}
\value{
prints .tif files to `out_path` of year 2 minus year 1
}
\description{
Given a directories of matching rasters from different years,
calculate change by subtracting year 1 from year 2.
}
