% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_change_report}
\alias{pred_change_report}
\title{predicted change report}
\usage{
pred_change_report(chng_path, year1_path, year2_path,
  permits = permits_13to18, lots = prc_lots,
  title = "pred_chng_table.csv")
}
\arguments{
\item{chng_path}{is a path for directory of change rasters output by `pred_change()`}

\item{year1_path}{is a character string path for directory of year 1 rasters}

\item{year2_path}{is a character string path for directory of year 2 rasters}

\item{permits}{is a list of permits issued by the county}

\item{lots}{is a spatial polygons object (maptaxlots)}

\item{title}{is the name given to the output csv file}
}
\value{
a csv to the working directory showing change per taxlot, and listing permits for each taxlot
}
\description{
Given a path to change rasters produced by `pred_change()`,
outputs a csv to the working directory of change per taxlot,
including permits for each taxlot.
}
