% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_cover}
\alias{pred_cover}
\title{plot predicted cover}
\usage{
pred_cover(
  polys,
  in_path = NULL,
  out_path = NULL,
  rgb_path = NULL,
  cir_path = NULL,
  buff = prc_buff
)
}
\arguments{
\item{in_path}{is a character string indicating the path to a directory to rasters}

\item{out_path}{is a character string indicating the path to the output directory}

\item{rgb_path}{is the path to 3-band rgb data}

\item{cir_path}{is the path to 3-band cir data}

\item{buff}{is the 50-ft riparian buffer polygon for `poly`}

\item{poly}{is a spatial polygons object (taxlot or sampling box)}
}
\value{
a raster plot of predicted cover masked by `poly`
}
\description{
Predicts cover extent within a polygon and returns the results as a raster plot
}
