% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_cover_report}
\alias{pred_cover_report}
\title{predicted cover report}
\usage{
pred_cover_report(in_path, taxlots = prc_lots)
}
\arguments{
\item{in_path}{is the character string path directory to rasters}

\item{taxlots}{is a polygon spatial object of maptaxlots}
}
\value{
prints a predicted cover table to the working directory
}
\description{
produces cover extent table from predicted cover rasters
}
