% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{pt_to_line}
\alias{pt_to_line}
\title{pt to line}
\usage{
pt_to_line(pt, line)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{line}{is SpatialLines object}
}
\value{
nearest point on line to `pt` among vertices in the line path
}
\description{
returns nearest point on line to `pt` among vertices in the line path
}
