% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_coords}
\alias{pull_coords}
\title{Convert SpatialPolygon to coordinate matrix}
\usage{
pull_coords(poly)
}
\arguments{
\item{poly}{is a SpatialPolygons object}
}
\value{
a matrix of coordinates representing the spatial object
}
\description{
Wrapper to convert polygons to coordinate matrices.
}
