% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quadcut}
\alias{quadcut}
\title{QUAD CUT RASTER}
\usage{
quadcut(ras)
}
\arguments{
\item{ras}{is a raster}
}
\value{
four equal parts of \code{ras} in a list
}
\description{
Bisects raster into four equal parts, returns the parts in a list.
}
