% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_samples}
\alias{random_samples}
\title{annual random monitoring samples}
\format{
an sf object with 54 multipolygons of 50 polygons each
}
\usage{
random_samples
}
\description{
a spatial object containing the 54 sampling boxes
for monitoring the priority riparian corridor
}
\keyword{datasets}
