% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_band}
\alias{read_band}
\title{Read band}
\usage{
read_band(poly, dir, band)
}
\arguments{
\item{poly}{is a spatial polygon object}

\item{dir}{is a directory of raster files}

\item{band}{is a number specifying the desired raster band}
}
\value{
the mean values of the raster band masked by the polygon
}
\description{
Returns the mean value of a given raster band across a polygon.
}
