% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_sample}
\alias{read_sample}
\title{Read sample}
\usage{
read_sample(poly, dir, band)
}
\arguments{
\item{poly}{is a spatial polygons object}

\item{dir}{is a path to a directory of raster files}

\item{band}{is the desired band of the raster to pull}
}
\value{
the mean values of the raster band across each slice of a sampling box
}
\description{
Get mean values of a raster band from each slice of a sampling box
}
