% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_box}
\alias{sample_box}
\title{sample box}
\usage{
sample_box(mid, left, right, id)
}
\arguments{
\item{mid}{is the centerline matrix output by centerline}

\item{left}{is the left bank matrix output by draw_bank}

\item{right}{is the right bank matrix output by draw_bank}

\item{id}{is a unique integer identifier for the sampling box}
}
\value{
a Polygons object (sampling box) containing 50 Polygon objects (slices)
}
\description{
Given points for the centerline, left and right bank, assembles the points
into 50 slices, and the slices into a spatial Polgyons object.
}
\seealso{
centerline

draw_bank
}
