% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_streams}
\alias{sample_streams}
\title{sample streams}
\usage{
sample_streams(n = 100, lots = prc_mtls, prc = prc_per,
  strms = prc_strms, buff = per_buff)
}
\arguments{
\item{n}{is the number of samples to take}

\item{lots}{is a polygon object representing sample extent}

\item{prc}{is the line path for perennial and fish-bearing streams in RR (NHD)}

\item{strms}{is the hydro-enforced drainage layer in RR (WSI)}
}
\value{
`n` points along `strms`
}
\description{
Generates `n` samples along `prc`, finds the closest points on `strms`
to each point along `prc`, returns `n` samples along `strms`.
}
