% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{segment}
\alias{segment}
\title{segment}
\usage{
segment(mat, steps)
}
\arguments{
\item{mat}{is the matrix of line coordinates}

\item{steps}{is the segment increment}
}
\value{
matrix of point coordinates for segments
}
\description{
Divide a line into even segments.
}
