% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{set_flag}
\alias{set_flag}
\title{set flag}
\usage{
set_flag(pt, mat, dist = 75 * 0.3048, us = F)
}
\arguments{
\item{pt}{is a coordinate pair (starting point)}

\item{mat}{is a matrix of coords representing a line with cols (x,y)}

\item{dist}{is a distance in meters}

\item{us}{is a logical boolean indicating whether distance is upstream or downstream}
}
\value{
the point distance `dist` from starting point `pt` along line `mat`
}
\description{
set flag
}
