% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{set_pt}
\alias{set_pt}
\title{set point}
\usage{
set_pt(pt, mat, dist, k = 0)
}
\arguments{
\item{pt}{is a coordinate pair (starting point)}

\item{mat}{is a matrix representing a line path with cols (x,y)}

\item{dist}{is a distance in meters}
}
\value{
the point distance `dist` along line `mat` from starting point `pt`
}
\description{
Given a starting point `pt` and a line `mat`,
returns a point distance `dist` from point `pt` along line `mat`
}
