% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_buff}
\alias{sf_buff}
\title{riparian buffer}
\format{
an sf multipolygon
}
\usage{
sf_buff
}
\description{
the 50-ft riparian buffer cropped to the priority riparian corridor
}
\keyword{datasets}
