% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_lots}
\alias{sf_lots}
\title{priority corridor taxlots}
\format{
an sf geometry collection of multipolygons with 630 obs. of 37 vars
}
\usage{
sf_lots
}
\description{
maptaxlots within the priority riparian corridor
}
\keyword{datasets}
