% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_perennial}
\alias{sf_perennial}
\title{perennial streams}
\format{
an sf geometry collection of multilinestrings with 127 obs of 38 vars
}
\usage{
sf_perennial
}
\description{
perennial stream layer from NHD
}
\keyword{datasets}
