% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{short_side}
\alias{short_side}
\title{short_side}
\usage{
short_side(x, y, circle)
}
\arguments{
\item{x}{is a point with elements c(x, y)}

\item{y}{is a point with elements c(x, y)}

\item{circle}{is a matrix of polygon coordinates}
}
\value{
the subset of `circle` containing the shorter path between `x` and `y`
}
\description{
Given two points on a polygon, return the shorter path between them.
}
