% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_fill}
\alias{simple_fill}
\title{Simple fill extent}
\usage{
simple_fill(poly, dir, band)
}
\arguments{
\item{poly}{is a spatial polygon object}

\item{dir}{is a path to a directory of rasters}

\item{band}{is the desired raster band}
}
\value{
the last raster to match in_extent
}
\description{
Returns last raster to match in_extent
}
