% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{snap_pt}
\alias{snap_pt}
\title{snap point}
\usage{
snap_pt(pt, mat)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{mat}{is a matrix with cols (x,y) and rows of points}
}
\value{
matrix `mat` with `pt` inserted by nearest neighbor
}
\description{
given a point along a stream path, and a matrix of points representing the stream path,
adds the point to the matrix of points and returns a new matrix
}
