% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spatialize}
\alias{spatialize}
\title{CONVERT XY COORDS TO SPATIALPOLYGONS}
\usage{
spatialize(mat, crs_ref)
}
\arguments{
\item{mat}{is a matrix of coords with cols (x,y)}

\item{crs_ref}{is a character string representing a coordinate projection for proj4string}
}
\value{
a SpatialPolygon defined by coords of mat in projection crs_ref
}
\description{
Given a matrix of coords with cols (x,y), and a character string representing a coordinate
projection, returns a SpatialPolygon defined by coords of \code{mat} in projection \code{crs_ref}.
}
