% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squash}
\alias{squash}
\title{Squash Characters Vectors}
\usage{
squash(vec)
}
\arguments{
\item{vec}{is a character vector}
}
\value{
vector contents as a string (length 1 char vector)
}
\description{
Takes a character vector and returns it as a string.
}
