% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_extent}
\alias{stack_extent}
\title{stack_extent}
\usage{
stack_extent(poly, dir)
}
\arguments{
\item{poly}{is a `SpatialPolygons` object}

\item{dir}{is a directory filepath character string}
}
\value{
merged rasters from `dir` over extent of `poly`
}
\description{
Given a `SpatialPolygons` object and a directory filepath character string,
returns merged rasters from `dir` over extent of `poly`.
Returns stack of all layers of og rasters
}
