% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{streamsnap}
\alias{streamsnap}
\title{SNAP POINT TO STREAM FLOWLINE}
\usage{
streamsnap(pt, mat)
}
\arguments{
\item{pt}{is a coordinate pair (numeric vector)}

\item{mat}{is a numeric matrix representing the stream path}
}
\value{
the point on the stream path closest to \code{pt}
}
\description{
SNAP POINT TO STREAM FLOWLINE
}
