% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thumbnails}
\alias{thumbnails}
\title{thumbnails}
\usage{
thumbnails(in_path, out_path, polys = random_samples, output = "tif")
}
\arguments{
\item{in_path}{is a character vector containing the path to the orthoimagery}

\item{out_path}{is a character vector specifying the output directory for the plots}

\item{polys}{is a SpatialPolygonsDataFrame object (the sample boxes shapefile)}

\item{output}{specifies type of output, 'tif' or 'png'}
}
\value{
saves rasters of extent `polys` to `out_path`
}
\description{
gets list of raster names from `in_path`
searches for rasters in extent of `polys`
mosaics rasters in extent together and crops to extent
saves raster in `out_path`
Output options:
}
\details{
* 'tif' makes a raster slightly larger than the target extent.
 * 'png' superimposes the polygon object and riparian buffer over the raster imagery.
}
