% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{traverse}
\alias{traverse}
\title{traverse}
\usage{
traverse(mat, dist, mats, thresh)
}
\arguments{
\item{mat}{is the current oriented stream}

\item{dist}{is the distance to travel}

\item{mats}{is the stream network}

\item{thresh}{is the connection threshold}
}
\value{
a list of c(next pt, distance remaining, orientation of travel)
}
\description{
Traverse down a stream network.
}
\seealso{
orient

walk
}
