% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{tri_length}
\alias{tri_length}
\title{triangle length}
\usage{
tri_length(pt, bear, dist, north = TRUE)
}
\arguments{
\item{pt}{is a coordinate pair}

\item{bear}{is a bearing in degrees}

\item{dist}{is a distance in meters}
}
\value{
coordinates of the point `dist` from `pt` along bearing `bear`
}
\description{
Returns the point a given distance and bearing from a coordinate pair.
}
