% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uturn}
\alias{uturn}
\title{uturn}
\usage{
uturn(bear)
}
\arguments{
\item{bear}{is a bearing in degrees}
}
\value{
a 180 degree change in bearing from `bear`
}
\description{
given a bearing, return a 180 degree change in bearing
bearings in degrees
}
