% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{walk}
\alias{walk}
\title{walk}
\usage{
walk(mat, dist)
}
\arguments{
\item{mat}{is a stream path}

\item{dist}{is the length to travel down the stream path}
}
\value{
a list c(new position, distance remaining, orientation of travel)
}
\description{
Traverse from top row to bottom row of matrix.  
If distance `dist` lies on matrix `mat`,  
return interpolated pt and remainder distance of zero.
If `dist` is farther than the end of `mat`, 
return end of mat as pt and remainder distance.
}
