% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap}
\alias{wrap}
\title{wrap}
\usage{
wrap(x, y)
}
\arguments{
\item{x}{is a matrix with cols c(x,y)}

\item{y}{is a matrix with cols c(x,y)}
}
\value{
a list of polgyons
}
\description{
Wrap two matrices of equal row number into box polygons.
}
