% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coord_lis}
\alias{coord_lis}
\title{COORDINATE LIST FROM SPATIALPOLYGONS}
\usage{
coord_lis(poly)
}
\arguments{
\item{poly}{is a \code{SpatialPolygons} object}
}
\value{
a list coordinates for vertices of polygons in \code{poly}
}
\description{
COORDINATE LIST FROM SPATIALPOLYGONS
}
