% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ext_box}
\alias{ext_box}
\title{box from extent}
\usage{
ext_box(ras, crs_ref)
}
\arguments{
\item{ras}{is a spatial object (raster)}
}
\value{
a Spatial Polygon drawn around the extent of `ras`
}
\description{
produce polygon representation of extent
}
