% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_extent}
\alias{fill_extent}
\title{fill_extent}
\usage{
fill_extent(poly, in_path, pad = 0.1)
}
\arguments{
\item{poly}{is a `sf` object}

\item{in_path}{is a directory filepath character string}

\item{pad}{is the padding factor}
}
\value{
merged rasters from `in_path` over extent of `poly`
}
\description{
Given a `sf` object and a directory filepath character string,
returns merged rasters from `in_path` over extent of `poly`.
}
\seealso{
before_and_after get_cols_4band plot_samples pred_cover1 sample_streams thumbnails
}
