% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lookup_lots}
\alias{lookup_lots}
\title{Look up MapTaxlot`}
\usage{
lookup_lots(so, lots)
}
\arguments{
\item{so}{is a spatial object (sampling boxes)}

\item{lots}{is an spdf of polygons (bc taxlots)}
}
\value{
the MapTaxlot numbers of each input polygon in a character vector
}
\description{
Given a list of polygons, returns character vector of all maptaxlot numbers
within each polygon.
}
