% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_crs}
\alias{match_crs}
\title{match_crs}
\usage{
match_crs(so, target)
}
\arguments{
\item{so}{is a spatial polygons objects}

\item{target}{is a spatial object with the target crs projection}
}
\value{
spatial object projected in the target crs
}
\description{
Make sure your extents overlap by converting polygons
into a common reference crs.
}
\seealso{
pred_cover1
}
