% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_samples}
\alias{plot_samples}
\title{Print Sample Plots}
\usage{
plot_samples(in_path, out_path, polys = samples, year, type = "random",
  method = "lm")
}
\arguments{
\item{in_path}{is a character vector containing the path to the orthoimagery}

\item{out_path}{is a character vector specifying the output directory for the plots}

\item{polys}{is a SpatialPolygonsDataFrame object (the sample boxes shapefile)}

\item{year}{is an integer specifying the year of orthographic survey}

\item{type}{is a character vector length 1 specifying sample type}

\item{method}{is a character vector specifying model type}
}
\value{
prints rgb plots of `polys` to `out_path` & a csv of predicted results
}
\description{
Superimposes sampling boxes over orthographic data.
Predicts the level of cover using specified `method`.
`method` options are c('lm', 'binom', 'lm3').
Default is 'lm', use 'lm3' for 3-band data.
}
