% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_samples}
\alias{plot_samples}
\title{Print Sample Plots}
\usage{
plot_samples(in_path, out_path, samples = samples, rgb_test = TRUE,
  ndvi_test = TRUE)
}
\arguments{
\item{in_path}{is a character vector containing the path to the orthoimagery}

\item{out_path}{is a character vector specifying the output directory for the plots}

\item{samples}{is a SpatialPolygonsDataFrame object (the sample boxes shapefile)}

\item{rgb}{is a logical indicator, prints rgb plots if true}

\item{ndvi}{is a logical indicator, prints ndvi plots if true}
}
\value{
prints an rgb and/or ndvi plot into the \code{out_path} directory
}
\description{
Given a shapefile of sampling boxes, and a path to a directory of 4-band ortho-imagery
prints rgb and/or ndvi plots of the sampling area with numbered boxes superimposed.
While the function prints both rgb and ndvi by default, the user can turn either off
by setting either the \code{rgb} or \code{ndvi} arguments to \code{FALSE}.
}
