% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_scores}
\alias{plot_scores}
\title{plot_scores}
\usage{
plot_scores(in_path, out_path, scores, polys = samples, title = "samples_")
}
\arguments{
\item{in_path}{is a character vector containing the path to the orthoimagery}

\item{out_path}{is a character vector specifying the output directory for the plots}

\item{scores}{is a data.table recording cover score observations}

\item{polys}{is a SpatialPolygonsDataFrame object (the sample boxes shapefile)}
}
\value{
prints rgb plots of `polys` to `out_path` with scores indicated by color
}
\description{
Superimposes sampling boxes over orthographic data.
Prints cover scores over each box.  
`scores` matches output format of plot_samples().  
`scores` must have same number of rows as `polys`, in the same order.
}
