% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{streamsnap}
\alias{streamsnap}
\title{snap point to stream}
\usage{
streamsnap(pt, mat)
}
\arguments{
\item{pt}{is a coordinate pair (numeric vector)}

\item{mat}{is a matrix of coords representing the stream path}
}
\value{
the point on the stream path closest to `pt`
}
\description{
Given a coordinate pair `pt` and a matrix of coords `mat`, 
returns the point on `mat` nearest `pt`.
}
