% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thumbnails}
\alias{thumbnails}
\title{thumbnails}
\usage{
thumbnails(in_path, out_path, polys = random_samples)
}
\arguments{
\item{in_path}{is a character vector containing the path to the orthoimagery}

\item{out_path}{is a character vector specifying the output directory for the plots}

\item{polys}{is a SpatialPolygonsDataFrame object (the sample boxes shapefile)}
}
\value{
saves rasters of extent `polys` to `out_path`
}
\description{
gets list of raster names from `in_path`
searches for rasters in extent of `polys`
mosaics rasters in extent together and crops to extent
saves raster in `out_path`
}
